# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UAV Sighting Reporting System API",
    "description": "Informant, agency, field, ingestion and analytics APIs for reporting.tw.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/maritime/alerts": {
      "get": {
        "tags": [
          "maritime"
        ],
        "summary": "Alerts",
        "operationId": "alerts_v1_maritime_alerts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "comma-separated; 'all' for every status",
              "default": "open,acknowledged",
              "title": "Status"
            },
            "description": "comma-separated; 'all' for every status"
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Min Score"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "default": 300,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Alerts V1 Maritime Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maritime/alerts/{alert_id}": {
      "get": {
        "tags": [
          "maritime"
        ],
        "summary": "Alert Detail",
        "operationId": "alert_detail_v1_maritime_alerts__alert_id__get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Alert Detail V1 Maritime Alerts  Alert Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maritime/alerts/{alert_id}/status": {
      "post": {
        "tags": [
          "maritime"
        ],
        "summary": "Decide",
        "operationId": "decide_v1_maritime_alerts__alert_id__status_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Decide V1 Maritime Alerts  Alert Id  Status Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maritime/alerts/{alert_id}/notes": {
      "post": {
        "tags": [
          "maritime"
        ],
        "summary": "Note",
        "operationId": "note_v1_maritime_alerts__alert_id__notes_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/uavr__api__maritime__NoteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Note V1 Maritime Alerts  Alert Id  Notes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maritime/alerts/{alert_id}/escalate": {
      "post": {
        "tags": [
          "maritime"
        ],
        "summary": "Escalate",
        "operationId": "escalate_v1_maritime_alerts__alert_id__escalate_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Escalate V1 Maritime Alerts  Alert Id  Escalate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maritime/settings": {
      "get": {
        "tags": [
          "maritime"
        ],
        "summary": "Get Settings ",
        "operationId": "get_settings__v1_maritime_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Get Settings  V1 Maritime Settings Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "maritime"
        ],
        "summary": "Put Settings",
        "operationId": "put_settings_v1_maritime_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Put Settings V1 Maritime Settings Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/maritime/evaluation": {
      "get": {
        "tags": [
          "maritime"
        ],
        "summary": "Evaluation",
        "operationId": "evaluation_v1_maritime_evaluation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Evaluation V1 Maritime Evaluation Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "maritime"
        ],
        "summary": "Evaluate Now",
        "description": "Compare the detectors on the labelled simulated tracks (optionally regenerate them first).",
        "operationId": "evaluate_now_v1_maritime_evaluation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_evaluate_now_v1_maritime_evaluation_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluate Now V1 Maritime Evaluation Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_evaluate_now_v1_maritime_evaluation_post": {
        "properties": {
          "resimulate": {
            "type": "boolean",
            "title": "Resimulate",
            "default": false
          }
        },
        "type": "object",
        "title": "Body_evaluate_now_v1_maritime_evaluation_post"
      },
...
      "DecisionIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "acknowledged",
              "false_alarm",
              "dismissed"
            ],
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "suppress_hours": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 720.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Suppress Hours"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DecisionIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "uavr__api__maritime__NoteIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "NoteIn"
      }
    }
  }
}